refactor and create customhook
