+++ src/hooks/useSuperHeroesData.js
@@ -1,0 +1,16 @@
+import { useQuery } from "react-query"
+import axios from 'axios'
+
+const fetchSuperHeroes = () => {
+  return axios.get('http://localhost:4000/superheroes')
+}
+
+export const useSuperHeroesData = () => {
+  return useQuery(
+    'super-heroes', 
+    fetchSuperHeroes, 
+    {
+      enabled: false,
+    }
+  )
+}
